requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.11.1
anyio==4.6.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.0.0
dask==2023.7.1
dask-sql==2023.8.0
Deprecated==1.2.14
distributed==2023.7.1
elastic-transport==8.15.1
elasticsearch==8.15.1
exceptiongroup==1.2.2
fastapi==0.115.0
Flask==3.0.3
flatten-json==0.1.14
fsspec==2024.9.0
ghp-import==2.1.0
googleapis-common-protos==1.65.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib-metadata==6.11.0
inflect==7.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
locket==1.0.0
Markdown==3.7
MarkupSafe==2.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-redoc-tag==0.1.0
mkdocs-swagger-ui-tag==0.6.10
more-itertools==10.5.0
msgpack==1.1.0
numpy==1.26.4
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
packaging==24.1
pandas==2.2.3
partd==1.4.2
pathspec==0.12.1
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==4.25.5
psutil==6.0.0
psycopg2==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyyaml_env_tag==0.1
requests==2.32.3
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
six==1.16.0
sniffio==1.3.1
soda-core==3.3.22
soda-core-pandas-dask==3.3.22
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
starlette==0.38.6
tabulate==0.9.0
tblib==3.0.0
toolz==1.0.0
tornado==6.4.1
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.31.1
watchdog==5.0.3
wcwidth==0.2.13
Werkzeug==3.0.4
wrapt==1.16.0
zict==3.0.0
zipp==3.20.2
